+++ src/input/edit.js
@@ -1,4 +1,4 @@
-import { useBlockProps } from '@wordpress/block-editor';
+import { RichText, useBlockProps } from '@wordpress/block-editor';
 import {
 	Flex,
 	FlexBlock,
@@ -69,12 +69,17 @@ export default function InputEdit( props ) {
 		value,
 		width,
 	};
+	const isButton = type === 'reset' || type === 'submit';
 	
 	blockProps.className += ' is-type-' + type;
 	
 	if ( type === 'hidden' ) {
 		elementProps.help = __( 'This input is hidden in the frontend.', 'form-block' );
 		elementProps.type = 'text';
+	}
+	
+	if ( isButton ) {
+		blockProps.className += ' wp-block-button';
 	}
 	
 	return (
@@ -92,11 +97,12 @@ export default function InputEdit( props ) {
 					
 					{ isAllowedAttribute( type, 'label' )
 						? <FlexBlock>
-							<TextControl
-								className="form-block__label-control"
+							<RichText
+								className="form-block__label"
 								onChange={ ( label ) => setAttributes( { label } ) }
 								placeholder={ __( 'Label', 'form-block' ) }
-								value={ label }
+								tagName="label"
+								value={ label || '' }
 							/>
 						</FlexBlock>
 						: null
@@ -119,11 +125,12 @@ export default function InputEdit( props ) {
 						? <Flex>
 							{ isAllowedAttribute( type, 'label' )
 								? <FlexBlock>
-									<TextControl
-										className="form-block__label-control"
+									<RichText
+										className="form-block__label"
 										onChange={ ( label ) => setAttributes( { label } ) }
 										placeholder={ __( 'Label', 'form-block' ) }
-										value={ label }
+										tagName="label"
+										value={ label || '' }
 									/>
 								</FlexBlock>
 								: null
@@ -144,6 +151,7 @@ export default function InputEdit( props ) {
 					}
 					
 					<TextControl
+						className={ isButton ? 'wp-block-button__link wp-element-button' : '' }
 						onChange={ ( value ) => setAttributes( { value } ) }
 						{ ...elementProps }
 					/>