+++ src/input/save.js
@@ -1,4 +1,5 @@
-import { useBlockProps } from '@wordpress/block-editor';
+import { RichText, useBlockProps } from '@wordpress/block-editor';
+
 import { getAllowedAttributes } from './html-data';
 
 export default function InputSave( props ) {
@@ -10,12 +11,13 @@ export default function InputSave( props ) {
 			type,
 		}
 	} = props;
-	const blockProps = useBlockProps.save( { className: 'form-block__element' } );
 	const allowedAttributes = getAllowedAttributes( type );
+	const blockProps = useBlockProps.save( { className: 'form-block__element' } );
 	let elementProps = {
 		name,
 		type,
 	};
+	const isButton = type === 'reset' || type === 'submit';
 	
 	for ( const allowedAttribute of allowedAttributes ) {
 		if ( allowedAttribute === 'label' ) {
@@ -27,6 +29,17 @@ export default function InputSave( props ) {
 	
 	blockProps.className += ' is-type-' + type;
 	
+	if ( isButton ) {
+		blockProps.className += ' wp-block-button';
+		
+		if ( elementProps.className ) {
+			elementProps.className += ' wp-block-button__link wp-element-button';
+		}
+		else {
+			elementProps.className = 'wp-block-button__link wp-element-button';
+		}
+	}
+	
 	return (
 		<div { ...blockProps }>
 			<input { ...elementProps } />
@@ -34,7 +47,11 @@ export default function InputSave( props ) {
 				? <label
 					className="form-block__label is-input-label"
 				>
-					<span className="form-block__label-content">{ label }</span>
+					<RichText.Content
+						className="form-block__label-content"
+						tagName="span"
+						value={ label }
+					/>
 					{ required ? <span className="is-required">*</span> : '' }
 				</label>
 				: null